requirements.txt
google-adk>=1.19.0
requests>=2.31.0

